rename DockedMegaMenu -> MegaMenu and clean up go code
