+++ public/app/core/components/AppChrome/MegaMenu/utils.ts
